requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Cython==0.29.34
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2
django-allauth==0.54.0
django-sslcommerz==1.0.0a0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
openai==0.27.7
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
pytesseract==0.3.10
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tqdm==4.65.0
typed-ast==1.5.5
urllib3==2.0.2
wrapt==1.11.2
yarl==1.9.2
